requirements.txt


fastapi
uvicorn
python-dotenv
requests
arxiv
groq


langchain>=0.1.20
chromadb==0.4.24
sentence-transformers
pypdf
arxiv
tiktoken


nginx
fastapi
uvicorn
python-dotenv
langchain-community
langchain-openai
chromadb
pypdf
sentence-transformers

fastapi
uvicorn
pydantic
